feature(log truncation): Allow logs to be truncated.
This allows dependency targets to specify whether or not their
logs(stdout, stderr) should be truncated before the first time they are
started up within the same process that depends-on runs.

If a depends-on target is used multiple times within the life of a
process, the log will only be truncated once.

Targets can add this by specifying the `truncateStdio` flag to `true` in
their configuration.